+++ src/main/java/roomescape/payment/domain/PaymentType.java
@@ -1,5 +1,5 @@
 package roomescape.payment.domain;
 
 public enum PaymentType {
-    TOSS
+    NORMAL
 }
