+++ src/main/java/roomescape/reservation/domain/dto/ReservationWithPaymentDto.java
@@ -1,11 +1,12 @@
 package roomescape.reservation.domain.dto;
 
-import java.time.LocalDate;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationStatus;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
 import roomescape.user.domain.User;
+
+import java.time.LocalDate;
 
 public record ReservationWithPaymentDto(LocalDate date,
                                         Long timeId,
@@ -13,7 +14,7 @@ public record ReservationWithPaymentDto(LocalDate date,
                                         String paymentKey,
                                         String orderId,
                                         int amount,
-                                        String paymentTypeText) {
+                                        String paymentType) {
 
     public Reservation toEntity(ReservationTime reservationTime, Theme theme, User user) {
         return Reservation.of(date, ReservationStatus.BOOKED, reservationTime, theme, user);
